Added \app\Session, added \views\_error, added \views\authUserList, added \views\userList, added \http\Controllers\userController, mod \views\layouts\main, mod \app\Application, mod \app\Request, mod \app\Response, mod \http\Models\User, mod \routes\user.routes.php
All of this to add further functionality to the program. this code makes use of the in place items to make sure things are working the way we would expect. tested my first POST request and so there's a function that doesn't really do much of anything but process and reply... But It works!